Des ia qui cherchent et qui bougent \o/

+++ server/inc/abstractions/maths.hh
@@ -5,7 +5,7 @@
 // Login   <aracthor@epitech.net>
 // 
 // Started on  Tue Oct 14 13:05:32 2014 
-// Last Update Fri Oct 24 16:30:13 2014 
+// Last Update Fri Nov  7 13:50:52 2014 
 //
 
 #ifndef MATHS_HH_
@@ -17,7 +17,7 @@
 # define MAX(a, b)	((a) > (b) ? (a) : (b))
 # define MIN(a, b)	((a) < (b) ? (a) : (b))
 
-# define RAD(a)		((a) * M_PI / 180.0f)
-# define DEG(a)		((a) / M_PI * 180.0f)
+# define RAD(a)		(((float)a) * M_PI / 180.0f)
+# define DEG(a)		(((float)a) / M_PI * 180.0f)
 
 #endif // !MATHS_HH_
